// src/services/-settings-get.schemas.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Shopping Platform API
 * Production-grade e-commerce backend
 * OpenAPI spec version: 1.0
 */
export interface UtilsResponse {
  code?: number;
  error?: string;
  message?: string;
  success?: boolean;
}

export type DtoSettingResponseValue = { [key: string]: unknown };

export interface DtoSettingResponse {
  created_at?: string;
  description?: string;
  key?: string;
  updated_at?: string;
  value?: DtoSettingResponseValue;
}

export type GetSettings200 = UtilsResponse & {
  data?: DtoSettingResponse[];
};